# edge_cloud_management_api/specification/openapi.yaml
...
openapi: 3.0.3
info:
  title: Open Exposure Gateway API
  version: 1.0.1-wip
  description: |
    Open Exposure Gateway API allows API consumers to manage the
    Life Cycle of an Application, Discover Edge Cloud Zones and request Network Resources.
    # Overview
    The reference scenario foresees a distributed Telco Edge Cloud where any
    Application Delevoper, known as an Application Provider, can host and
    deploy their application according to their specifications and operational
    criteria (e.g. within an specific geographical zone for data protection
    purposes, ensure a minimum QoS for the application clients, etc).
    Through Telco Edge Cloud services Developers around the globe can be
    benefit from the traditional Cloud strengths but expertise and advantages
    of the Mobile Network Operators offering to their users an evolved
    experience for XR, V2X, Holographic and other new services.

    # Introduction
    The Edge Application Management API provides capabilities for lifecycle
    management of application, instances and edge cloud zone discovery.
    Lifecycle Management allows Application Provider to onboard
    their application to the Edge Cloud Platform which do bookkeeping,
    resource validation and other pre-deployment operations.
    Application details can contain components network specification,
    package type (QCOW2, OVA, CONTAINER, HELM), operating system details and
...
    Once the application is available on the Edge Cloud
...

...

...

...

...

...

...
servers:
...
paths:
  /apps:
    post:
      # security:
      #   - openId:
      #       - edge-application-management:apps:write
      tags:
        - Application
      summary: Submit application metadata to the Edge Cloud Provider.
      description: |
        Contains the information about the application to be
        instantiated in the Edge Cloud
      operationId: edge_cloud_management_api.controllers.app_controllers.submit_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      requestBody:
        description: |
          The Application Provider request contains mandatory
          criteria (e.g. required CPU, memory, storage, bandwidth) and
          optional parameters.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppManifest"
            examples:
              KubernetesExample:
                summary: Example for Kubernetes app
                value:
                  appId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  name: nginx_web_app
                  appProvider: nginx_inc
                  version: 1.0.0
                  packageType: HELM
                  operatingSystem:
                    architecture: x86_64
                    family: UBUNTU
                    version: OS_VERSION_UBUNTU_2204_LTS
                    license: OS_LICENSE_TYPE_FREE
                  appRepo:
                    type: PRIVATEREPO
                    imagePath: "https://charts.bitnami.com/bitnami/nginx:1.25.2"
                    userName: helm-user
                    credentials: secure-token
                    authType: DOCKER
                    checksum: sha256:4d67e4c1a4b4c8...
                  requiredResources:
                    infraKind: kubernetes
                    applicationResources:
                      cpuPool:
                        numCPU: 2
                        memory: 4096
                        topology:
                          minNumberOfNodes: 3
                          minNodeCpu: 2
                          minNodeMemory: 4096
                      gpuPool:
                        numCPU: 4
                        memory: 8192
                        gpuMemory: 32
                        topology:
                          minNumberOfNodes: 2
                          minNodeCpu: 4
                          minNodeMemory: 8192
                          minNodeGpuMemory: 16
                    isStandalone: false
                    version: "1.29"
                    additionalStorage: 100GB
                    networking:
                      primaryNetwork:
                        provider: cilium
                        version: "1.13"
                      additionalNetworks:
                        - name: net1
                          interfaceType: vfio-pci
                        - name: backend-net
                          interfaceType: netdevice
                        - name: db-net
                          interfaceType: interface
                    addons:
                      monitoring: true
                      ingress: true
                  componentSpec:
                    - componentName: nginx_server
                      networkInterfaces:
                        - interfaceId: eth0
                          protocol: TCP
                          port: 80
                          visibilityType: VISIBILITY_EXTERNAL
                        - interfaceId: eth1
                          protocol: TCP
                          port: 443
                          visibilityType: VISIBILITY_EXTERNAL

        required: true
      responses:
        "201":
          description: Application created successfully
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmittedApp"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "App already exists"
        "500":
          $ref: "#/components/responses/500"
        "501":
          $ref: "#/components/responses/501"
        "503":
          $ref: "#/components/responses/503"
    get:
      # security:
      #   - openId:
      #       - edge-application-management:apps:read
      tags:
        - Application
      summary: Retrieve a list of existing Applications
      description: |
        Get the list of all existing Application definitions from the
        Edge Cloud Provider that the user has permission to view.
      operationId: edge_cloud_management_api.controllers.app_controllers.get_apps
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      responses:
        "200":
          description: List of existing applications
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppManifest"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /apps/{appId}:
    get:
      # security:
      #   - openId:
      #       - edge-application-management:apps:read
      tags:
        - Application
      summary: Retrieve the information of an Application
      description: |
        Ask the Edge Cloud Provider the information for a given application
      operationId: edge_cloud_management_api.controllers.app_controllers.get_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          description: |
            A globally unique identifier associated with the
            application.
            Edge Cloud Provider generates this identifier when the application
            is submitted.
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AppId"
      responses:
        "200":
          description: Information of Application
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: object
                properties:
                  appManifest:
                    $ref: "#/components/schemas/AppManifest"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
    delete:
      # security:
      #   - openId:
      #       - edge-application-management:apps:delete
      tags:
        - Application
      summary: |
        Delete an Application from an Edge Cloud Provider
      description: Delete all the information and content
        related to an Application
      operationId: edge_cloud_management_api.controllers.app_controllers.delete_app
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          in: path
          description: |
            Identificator of the application to be
            deleted provided by the Edge Cloud Provider
            once the submission was successful
          required: true
          schema:
            $ref: "#/components/schemas/AppId"
      responses:
        "202":
          description: Request accepted
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
        "204":
          description: App deleted
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "App with a running application instance
                  cannot be deleted"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /appinstances:
    post:
      # security:
      #   - openId:
      #       - edge-application-management:instances:write
      tags:
        - Application
      summary: Instantiation of an Application
      description: |
        Ask the Edge Cloud Platform to instantiate an application to one
        or several Edge Cloud Zones with an Application as an input and an
        Application Instance as the output.
      operationId: edge_cloud_management_api.controllers.app_controllers.create_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
      requestBody:
        description: |
          Information about the application and where to deploy it.
        content:
          application/json:
            schema:
              type: object
              required:
                # - name
                - appId
                - appZones
                # - edgeCloudZoneId
              properties:
                # name: 
                  # $ref: '#/components/schemas/AppInstanceName'
                appId:
                  $ref: "#/components/schemas/AppId"
                # edgeCloudZoneId:
                #   $ref: "#/components/schemas/EdgeCloudZoneId"
                # kubernetesClusterRef:
                #   $ref: "#/components/schemas/KubernetesClusterRef"
                appZones:
                  $ref: "#/components/schemas/EdgeCloudZones"
        required: true
      responses:
        "202":
          description: Application instantiation accepted
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
            Location:
              description: Contains the URI of the newly created application.
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  appInstances:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppInstanceInfo"
                minItems: 1
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
              example:
                status: 409
                code: CONFLICT
                message: "Application already instantiated in the given
                  Edge Cloud Zone or Edge Cloud Region"
        "500":
          $ref: "#/components/responses/500"
        "501":
          $ref: "#/components/responses/501"
        "503":
          $ref: "#/components/responses/503"
    get:
      # security:
      #   - openId:
      #       - edge-application-management:instances:read
      tags:
        - Application
      summary: Retrieve the information of Application Instances for a given App
      description: |
        Ask the Edge Cloud Provider the information of the instances for a
        given application
      operationId: edge_cloud_management_api.controllers.app_controllers.get_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appId
          description: |
            A globally unique identifier associated with
            the application.
            Edge Cloud Provider generates this identifier when the
            application is submitted.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AppId"
        - name: appInstanceId
          description: |
            A globally unique identifier associated with a running
            instance of an application within an specific Edge Cloud Zone.
            Edge Cloud Provider generates this identifier.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AppInstanceId"
        - name: region
          description: |
            Human readable name of the geographical Edge Cloud Region of
            the Edge Cloud. Defined by the Edge Cloud Provider.
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EdgeCloudRegion"
      responses:
        "200":
          description: Information of Application Instances
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
          content:
            application/json:
              schema:
                type: object
                properties:
                  appInstanceInfo:
                    type: array
                    items:
                      $ref: "#/components/schemas/AppInstanceInfo"
                minItems: 1
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
  /appinstances/{appInstanceId}:
    delete:
      # security:
      #   - openId:
      #       - edge-application-management:instances:delete
      tags:
        - Application
      summary: Terminate an Application Instance
      description: |
        Terminate a running instance of an application within
        an Edge Cloud Zone
      operationId: edge_cloud_management_api.controllers.app_controllers.delete_app_instance
      parameters:
        - $ref: "#/components/parameters/x-correlator"
        - name: appInstanceId
          in: path
          description: |
            Identificator of the specific application instance
            that will be terminated
          required: true
          schema:
            $ref: "#/components/schemas/AppInstanceId"
      responses:
        "202":
          description: |
            Request accepted to be processed. It applies for async
            deletion process
          headers:
            x-correlator:
              $ref: "#/components/headers/x-correlator"
        "204":
          description: Application Instance Deleted
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
...
components:
...
  parameters:
    x-correlator:
      name: x-correlator
      in: header
      description: |
        Correlation id for the different services
      schema:
        type: string
  headers:
    x-correlator:
      description: |
        Correlation id for the different services
      required: false
      schema:
        type: string
        format: uuid
  schemas:
...
    AppManifest:
      description: |
        Application information and requirements provided by the
        Application Provider
      type: object
      properties:
        appId:
          $ref: "#/components/schemas/AppId"
        name:
          type: string
          # pattern: ^[A-Za-z][A-Za-z0-9_]{1,63}$
          description: Name of the application.
        appProvider:
          $ref: "#/components/schemas/AppProvider"
        version:
          type: string
          description: Application version information
        packageType:
          description: Format of the application image package
          type: string
          enum:
            - QCOW2
            - OVA
            - CONTAINER
            - HELM
        operatingSystem:
          $ref: "#/components/schemas/OperatingSystem"
        appRepo:
          description: |
            Repository where Application Provider stores the application image
          type: object
          required:
            - type
            - imagePath
          properties:
            type:
              type: string
              enum:
                - PRIVATEREPO
                - PUBLICREPO
              description: |
                Application repository and image URI information.
                PUBLICREPO is used of public urls like github, helm repo etc.
                PRIVATEREPO  is used for private repo managed by the application
                developer. Private repo can be accessed by using the app
                developer provided userName and password. Password is
                recommended to be the personal access token created by developer
                e.g. in Github repo.
            imagePath:
              $ref: "#/components/schemas/Uri"
            userName:
              type: string
              description: |
                Username to acces the Helm chart, docker-compose
                file or VM image repository
            credentials:
              type: string
              maxLength: 128
              description: |
                Password or personal access token created by
                developer to acces the app repository. API users can generate
                a personal access token e.g. docker clients to use them as
                password.
            authType:
              type: string
              enum:
                - DOCKER
                - HTTP_BASIC
                - HTTP_BEARER
                - NONE
              description: |
                The credentials can also be formatted as a Basic
                auth or Bearer auth in HTTP "Authorization" header.
            checksum:
              type: string
              description: |
                MD5 checksum for VM and file-based images, sha256
                digest for containers
        requiredResources:
          $ref: "#/components/schemas/RequiredResources"
        componentSpec:
          description: |
            Information defined in "appRepo" point to the application
            descriptor e.g. Helm chart, docker-compose yaml file etc.
            The descriptor may contain one or more containers and their
            associated meta-data. A component refers to additional details
            about these containers to expose the instances of the containers
            to external client applications. App provider can define one or
            more components (via the associated network port) in componentSpec
            corresponding to the containers in helm charts or docker-compose
            yaml file as part of app descriptors.
          type: array
          items:
            type: object
            required:
              - componentName
              - networkInterfaces
            properties:
              componentName:
                type: string
                description: Component name must be unique with an application
              networkInterfaces:
                description: |
                  Each application component exposes some ports
                  either for external users or for inter component
                  communication.
                  Application provider is required to specify which ports are
                  to be exposed and the type of traffic that will flow through
                  these ports.The underlying platform may assign a dynamic port
                  against the "extPort" that the application clients will use
                  to connect with edge application instance.
                type: array
                items:
                  type: object
                  required:
                    - interfaceId
                    - protocol
                    - port
                    - visibilityType
                  properties:
                    interfaceId:
                      type: string
                      # pattern: ^[A-Za-z][A-Za-z0-9_]{3,31}$
                      description: |
                        Each Port and corresponding traffic protocol
                        exposed by the component is identified by a name.
                        Application client on user device requires this to
                        uniquley idenify the interface.
                    protocol:
                      type: string
                      enum:
                        - TCP
                        - UDP
                        - ANY
                      description: |
                        Defines the IP transport communication
                        protocol i.e., TCP, UDP or ANY
                    port:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 65535
                      description: |
                        Port number exposed by the component.
                        Edge Cloud Provider may generate a dynamic port
                        towards the component instance which forwards
                        external traffic to the component port.
                    visibilityType:
                      description: |
                        Defines whether the interface is exposed
                        to outer world or not i.e., external, or internal.
                        If this is set to "external", then it is  exposed
                        to external applications otherwise it is exposed
                        internally to edge application components within
                        edge cloud. When exposed to external world,
                        an external dynamic port is assigned for UC traffic
                        and mapped to the extPort
                      type: string
                      enum:
                        - VISIBILITY_EXTERNAL
                        - VISIBILITY_INTERNAL
                minItems: 1

      required:
        - name
        - version
        # - appProvider
        - packageType
        - appRepo
        # - requiredResources
        - componentSpec
...